+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/JapanSearchKotlinExampleActivity.kt
@@ -1,0 +1,79 @@
+package com.mapbox.search.sample.api
+
+import android.app.Activity
+import android.os.Bundle
+import android.util.Log
+import com.mapbox.search.ApiType
+import com.mapbox.search.Country
+import com.mapbox.search.Language
+import com.mapbox.search.MapboxSearchSdk
+import com.mapbox.search.ResponseInfo
+import com.mapbox.search.SearchEngine
+import com.mapbox.search.SearchEngineSettings
+import com.mapbox.search.SearchOptions
+import com.mapbox.search.SearchRequestTask
+import com.mapbox.search.SearchSelectionCallback
+import com.mapbox.search.result.SearchResult
+import com.mapbox.search.result.SearchSuggestion
+import com.mapbox.search.sample.BuildConfig
+
+class JapanSearchKotlinExampleActivity : Activity() {
+
+    private lateinit var searchEngine: SearchEngine
+    private lateinit var searchRequestTask: SearchRequestTask
+
+    private val searchCallback = object : SearchSelectionCallback {
+
+        override fun onSuggestions(suggestions: List<SearchSuggestion>, responseInfo: ResponseInfo) {
+            if (suggestions.isEmpty()) {
+                Log.i("SearchApiExample", "No suggestions found")
+            } else {
+                Log.i("SearchApiExample", "Search suggestions: $suggestions.\nSelecting first suggestion...")
+                searchRequestTask = searchEngine.select(suggestions.first(), this)
+            }
+        }
+
+        override fun onResult(
+            suggestion: SearchSuggestion,
+            result: SearchResult,
+            responseInfo: ResponseInfo
+        ) {
+            Log.i("SearchApiExample", "Search result: $result")
+        }
+
+        override fun onCategoryResult(
+            suggestion: SearchSuggestion,
+            results: List<SearchResult>,
+            responseInfo: ResponseInfo
+        ) {
+            Log.i("SearchApiExample", "Category search results: $results")
+        }
+
+        override fun onError(e: Exception) {
+            Log.i("SearchApiExample", "Search error", e)
+        }
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        searchEngine = MapboxSearchSdk.createSearchEngineWithBuiltInDataProviders(
+            ApiType.SBS,
+            SearchEngineSettings(BuildConfig.MAPBOX_API_TOKEN)
+        )
+
+        searchRequestTask = searchEngine.search(
+            "東京",
+            SearchOptions(
+                countries = listOf(Country.JAPAN),
+                languages = listOf(Language.JAPANESE),
+            ),
+            searchCallback
+        )
+    }
+
+    override fun onDestroy() {
+        searchRequestTask.cancel()
+        super.onDestroy()
+    }
+}